Fixed type conversion.

+++ include/arctk/torus/sim.inl
@@ -359,7 +359,7 @@ namespace arc //! arctk namespace
 
             for (size_t i = 0; i < _lights.size(); ++i)
             {
-                _num_phot_per_light[i] = _num_phot * (_lights[i].second / total_ratio);
+                _num_phot_per_light[i] = static_cast<unsigned long int>(static_cast<double>(_num_phot) * (_lights[i].second / total_ratio));
             }
 
             const unsigned long int excess_phot = math::container::sum(_num_phot_per_light) - _num_phot;
